requirements.txt
fastapi
uvicorn
redis
sqlalchemy
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
